Removed unused trigger procedures
On Waag's master branch, when a node or node data is updated the
bounding box of the layer on which that node or node data is on is
updated. This is done by an after insert trigger. Our fork does not
calculate the bounding box at any point (it's left NULL).

Waag could not include Manchester's endpoint in their discovery service
because our layers do not have a bounding box.

I reinstated the triggers and attempted to rebuild the database. This
caused an unacceptable performance decrease. As a result, I've removed
the triggers and will instead explicitly call the bounding box update
procedure after a (bulk) insert is complete.